bot/requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
asyncio-mqtt==0.16.2
aiohttp==3.9.1